+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Show progress feedback ("Creating PR for... #number") when opening PRs for the first time
+
 ## [1.0.0-beta.2] - 2026-01-12
 
 ### Added